Add ability to specify driver and dialect through env vars
We need this since mssql requires `driverClassName` to be set to
`com.microsoft.sqlserver.jdbc.SQLServerDriver` and the dialect may be
needed to identify exact dialect of protocol that teh DB uses.

Note: This change breaks implementations that do not specify this
variable so we need to move this to code config part before merging to
master.