Temp fix for CVE-2021-23840 - remove openssl from image
This change removes the openssl package from the Pet Store
Docker image as a temporary workaround for CVE-2021-23840.

The assumption is that we don't require OpenSSL in the final
Pet Store application Docker image.

It's expected that the Maven and OpenJDK images upon which the
Pet Store app is based will eventually be upgraded to include
openSSL version 1.1.1j or later. These newer versions of OpenSSL
contain fixes for CVE-2021-23840 and CVE-2021-23841.